ID mapping for dashboards
