+++ src/data/SideBarData.ts
@@ -19,8 +19,18 @@ export const data = [
 		id: 0,
 		name: "Dashboard",
 		icon: MdDashboard,
-		subtopics: [],
-		path: "/home",
+		subtopics: [
+			{id: 2000, name: "Default Dashboard"},
+			{id: 2001, name: "Financial/Accounting Dashboard"},
+			{id: 2002, name: "Customer Management Dashboard"},
+			{id: 2003, name: "Operations/Field Dashboard"},
+			{id: 2004, name: "Analytics Dashboard"},
+			{id: 2005, name: "Solar Operations Dashboard"},
+			{id: 2006, name: "Collections & Payments Dashboard"},
+			{id: 2007, name: "Executive/KPI Dashboard"},
+			{id: 2008, name: "Inventory & Procurement Dashboard"},
+		],
+		path: "/dashboard",
 	},
 	{
 		id: 1,
